refactor(engine): split GpuReplay into replay/ submodules (C1 LOC-cap)
replay.rs (1617 LOC, over the C1 <1500 cap) is split move-only into a
replay/ directory:

  replay/mod.rs    (603)  GpuReplay struct + dispatch core: new /
                          update_viewport / submit / reintegrate_offscreen_content
  replay/flush.rs  (1052) the canonical five-phase flush machinery:
                          flush_segment + the five per-bucket flush helpers
                          (instanced batches, gradients, tessellated geometry,
                          cached/external images) + the four flush_texture_batch*
                          blend variants

Both files are now under the C1 1500-LOC cap. flush.rs is a descendant
module of replay, so its inherent `impl GpuReplay` methods retain access to
GpuReplay's private fields. The cut is at flush_segment: submit calls only
flush_segment + flush_texture_batch_premultiplied, both already
`pub(in crate::wgpu)`, and the five private bucket helpers are called only
within flush.rs — so the split needs ZERO visibility changes. The sibling
`ssaa.rs` `impl GpuReplay` block (the existing multi-file-impl precedent) is
untouched and still reaches the pub(super) fields.

Inline `super::` paths in the moved methods (instancing / multi_draw /
texture_cache — the only wgpu-level refs) are rebased to `super::super::` for
the one-level-deeper module; use blocks are split to each file's actual needs.

Pure move: zero behavior change, proven by the full enable-wgpu-tests GPU
readback suite (443 passed, 0 failed) staying bit-identical.

Doc references to the now-deleted replay.rs that point at relocated flush code
are updated; the generic descriptive comments are deferred to the
doc-reconciliation pass.
